requirements.txt
anthropic
langgraph
langchain-anthropic
streamlit
pydantic>=2.0
python-dotenv
pytest
